Move BuildConfiguration to container package
